+++ python/ray/rllib/a3c/shared_model_lstm.py
@@ -4,19 +4,26 @@ from __future__ import print_function
 
 import tensorflow as tf
 from ray.rllib.models.misc import linear, normc_initializer
+from ray.rllib.models.catalog import ModelCatalog
 from ray.rllib.a3c.policy import Policy
 
-from ray.rllib.models.catalog import ModelCatalog
+from ray.rllib.models.lstm import LSTM
 
 
-class SharedModel(Policy):
+class SharedModelLSTM(Policy):
+
     def __init__(self, ob_space, ac_space, **kwargs):
-        super(SharedModel, self).__init__(ob_space, ac_space, **kwargs)
+        super(SharedModelLSTM, self).__init__(ob_space, ac_space, **kwargs)
 
     def setup_graph(self, ob_space, ac_space):
         self.x = tf.placeholder(tf.float32, [None] + list(ob_space))
         dist_class, self.logit_dim = ModelCatalog.get_action_dist(ac_space)
-        self._model = ModelCatalog.ConvolutionalNetwork(self.x, self.logit_dim)
+        self._model = LSTM(self.x, self.logit_dim, {})
+
+        self.state_init = self._model.state_init
+        self.state_in = self._model.state_in
+        self.state_out = self._model.state_out
+
         self.logits = self._model.outputs
         self.curr_dist = dist_class(self.logits)
         # with tf.variable_scope("vf"):
@@ -33,16 +40,22 @@ class SharedModel(Policy):
             trainable=False)
 
     def get_gradients(self, batch):
-        info = {}
+        """Computing the gradient is actually model-dependent.
+
+        The LSTM needs its hidden states in order to compute the gradient
+        accurately.
+        """
         feed_dict = {
             self.x: batch.si,
             self.ac: batch.a,
             self.adv: batch.adv,
             self.r: batch.r,
+            self.state_in[0]: batch.features[0],
+            self.state_in[1]: batch.features[1]
         }
-        self.grads = [g for g in self.grads if g is not None]
+        info = {}
         self.local_steps += 1
-        if self.summarize:
+        if self.summarize and self.local_steps % 10 == 0:
             grad, summ = self.sess.run([self.grads, self.summary_op],
                                        feed_dict=feed_dict)
             info['summary'] = summ
@@ -50,13 +63,20 @@ class SharedModel(Policy):
             grad = self.sess.run(self.grads, feed_dict=feed_dict)
         return grad, info
 
-    def compute_actions(self, ob, *args):
-        action, vf = self.sess.run([self.sample, self.vf],
-                                   {self.x: [ob]})
-        return action[0], vf
+    def compute_actions(self, ob, c, h):
+        output = self.sess.run([self.sample, self.vf] + self.state_out,
+                               {self.x: [ob],
+                                self.state_in[0]: c,
+                                self.state_in[1]: h})
+        output = list(output)
+        output[0] = output[0][0]
+        return output
 
-    def value(self, ob, *args):
-        return self.sess.run(self.vf, {self.x: [ob]})[0]
+    def value(self, ob, c, h):
+        # process_rollout is very non-intuitive due to value being a float
+        return self.sess.run(self.vf, {self.x: [ob],
+                                       self.state_in[0]: c,
+                                       self.state_in[1]: h})[0]
 
     def get_initial_features(self):
-        return []
+        return self.state_init
