+++ python/ray/rllib/models/fcnet.py
@@ -8,15 +8,7 @@ import tensorflow.contrib.slim as slim
 import numpy as np
 
 from ray.rllib.models.model import Model
-
-
-def normc_initializer(std=1.0):
-    def _initializer(shape, dtype=None, partition_info=None):
-        out = np.random.randn(*shape).astype(np.float32)
-        out *= std / np.sqrt(np.square(out).sum(axis=0, keepdims=True))
-        return tf.constant(out)
-    return _initializer
-
+from ray.rllib.models.misc import normc_initializer
 
 class FullyConnectedNetwork(Model):
     """Generic fully connected network."""
